Refactor: move version.h from source_main to source_base
version.h is a leaf header (a single VERSION macro) with no dependencies,
but it lived in source_main (the top-level entry layer). This forced
source_io to include upward into source_main (edge source_io -> source_main),
a reverse dependency that violates the intended module layering.

Move it to source_base (L0 infrastructure), which every module may depend
on, and update the 6 include sites. No behavior change; header-only, so no
CMake changes are needed (resolved via the existing source/ include root).

This cuts the source_io -> source_main reverse edge.
