+++ source/source_io/module_json/test/para_json_test.cpp
@@ -7,7 +7,7 @@
 #include "source_io/module_json/readin_info.h"
 #include "source_io/module_parameter/parameter.h"
 #include "source_io/module_json/para_json.h"
-#include "source_main/version.h"
+#include "source_base/version.h"
 #undef private
 /************************************************
  *  unit test of json output module
